Запуск счетчика при выборе страницы призов
